Add config flags for sync source and destination
With this change, it is now possible to personalize the names of the
source and the destination from the rclone config file, instead of
relying on the hardcoded names "source:" and "destination:".

Besides using environment variables, the configuration can be provided
as an scheduled event input. The configuration names remain the same.